+++ plugins/subscriptions/src/subscriptions.js
@@ -1,0 +1,2 @@
+export const subscriptions = new Map()
+export const patternSubscriptions = new Map()
